fix initiate resumable upload tests
