+++ tests/unit/test_client.py
@@ -16,7 +16,6 @@ import copy
 import collections
 import datetime
 import decimal
-import email
 import gzip
 import http.client
 import io
@@ -28,12 +27,9 @@ import unittest
 from unittest import mock
 import warnings
 
-import freezegun
 import packaging
 import pytest
 import requests
-
-import google.api
 
 
 try:
@@ -58,8 +54,6 @@ from google.api_core import client_info
 import google.cloud._helpers
 from google.cloud import bigquery
 
-from google.cloud.bigquery import job as bqjob
-import google.cloud.bigquery._job_helpers
 from google.cloud.bigquery.dataset import DatasetReference, Dataset
 from google.cloud.bigquery.enums import UpdateMode, DatasetView
 from google.cloud.bigquery import exceptions
@@ -312,31 +306,6 @@ class TestClient(unittest.TestCase):
         _, kwargs = http.request.call_args
         headers = kwargs["headers"]
         assert headers["x-goog-request-reason"] == "because-friday"
-
-    def test__call_api_applying_custom_retry_on_timeout(self, global_time_lock):
-        from concurrent.futures import TimeoutError
-        from google.cloud.bigquery.retry import DEFAULT_RETRY
-
-        creds = _make_credentials()
-        client = self._make_one(project=self.PROJECT, credentials=creds)
-
-        api_request_patcher = mock.patch.object(
-            client._connection,
-            "api_request",
-            side_effect=[TimeoutError, "result"],
-        )
-        retry = DEFAULT_RETRY.with_deadline(1).with_predicate(
-            lambda exc: isinstance(exc, TimeoutError)
-        )
-
-        with api_request_patcher as fake_api_request:
-            result = client._call_api(retry, foo="bar")
-
-        self.assertEqual(result, "result")
-        self.assertEqual(
-            fake_api_request.call_args_list,
-            [mock.call(foo="bar"), mock.call(foo="bar")],  # was retried once
-        )
 
     def test__call_api_span_creator_not_called(self):
         from concurrent.futures import TimeoutError
@@ -643,48 +612,6 @@ class TestClient(unittest.TestCase):
             method="GET", path=path, timeout=DEFAULT_TIMEOUT
         )
         self.assertEqual(service_account_email, email)
-
-    def test_get_service_account_email_w_custom_retry(self, global_time_lock):
-        from google.cloud.bigquery.retry import DEFAULT_RETRY
-
-        api_path = "/projects/{}/serviceAccount".format(self.PROJECT)
-        creds = _make_credentials()
-        http = object()
-        client = self._make_one(project=self.PROJECT, credentials=creds, _http=http)
-
-        resource = {
-            "kind": "bigquery#getServiceAccountResponse",
-            "email": "bq-123@bigquery-encryption.iam.gserviceaccount.com",
-        }
-        api_request_patcher = mock.patch.object(
-            client._connection,
-            "api_request",
-            side_effect=[ValueError, resource],
-        )
-
-        retry = DEFAULT_RETRY.with_deadline(1).with_predicate(
-            lambda exc: isinstance(exc, ValueError)
-        )
-
-        with api_request_patcher as fake_api_request:
-            with mock.patch(
-                "google.cloud.bigquery.opentelemetry_tracing._get_final_span_attributes"
-            ) as final_attributes:
-                service_account_email = client.get_service_account_email(
-                    retry=retry, timeout=7.5
-                )
-
-        final_attributes.assert_called_once_with({"path": api_path}, client, None)
-        self.assertEqual(
-            service_account_email, "bq-123@bigquery-encryption.iam.gserviceaccount.com"
-        )
-        self.assertEqual(
-            fake_api_request.call_args_list,
-            [
-                mock.call(method="GET", path=api_path, timeout=7.5),
-                mock.call(method="GET", path=api_path, timeout=7.5),  # was retried once
-            ],
-        )
 
     def test_dataset_with_specified_project(self):
         from google.cloud.bigquery.dataset import DatasetReference
@@ -3865,159 +3792,6 @@ class TestClient(unittest.TestCase):
         fake_transport.request.return_value = fake_response
         return fake_transport
 
-    def _initiate_resumable_upload_helper(self, num_retries=None, mtls=False):
-        from google.resumable_media.requests import ResumableUpload
-        from google.cloud.bigquery.client import _DEFAULT_CHUNKSIZE
-        from google.cloud.bigquery.client import _GENERIC_CONTENT_TYPE
-        from google.cloud.bigquery.client import _get_upload_headers
-        from google.cloud.bigquery.job import LoadJob
-        from google.cloud.bigquery.job import LoadJobConfig
-        from google.cloud.bigquery.job import SourceFormat
-
-        # Create mocks to be checked for doing transport.
-        resumable_url = "http://test.invalid?upload_id=hey-you"
-        response_headers = {"location": resumable_url}
-        fake_transport = self._mock_transport(http.client.OK, response_headers)
-        client = self._make_one(project=self.PROJECT, _http=fake_transport)
-        conn = client._connection = make_connection()
-        if mtls:
-            conn.get_api_base_url_for_mtls = mock.Mock(return_value="https://foo.mtls")
-
-        # Create some mock arguments and call the method under test.
-        data = b"goodbye gudbi gootbee"
-        stream = io.BytesIO(data)
-        config = LoadJobConfig()
-        config.source_format = SourceFormat.CSV
-        job = LoadJob(None, None, self.TABLE_REF, client, job_config=config)
-        metadata = job.to_api_repr()
-        upload, transport = client._initiate_resumable_upload(
-            stream, metadata, num_retries, None
-        )
-
-        # Check the returned values.
-        self.assertIsInstance(upload, ResumableUpload)
-
-        host_name = "https://foo.mtls" if mtls else "https://bigquery.googleapis.com"
-        upload_url = (
-            f"{host_name}/upload/bigquery/v2/projects/{self.PROJECT}"
-            "/jobs?uploadType=resumable"
-        )
-        self.assertEqual(upload.upload_url, upload_url)
-        expected_headers = _get_upload_headers(conn.user_agent)
-        self.assertEqual(upload._headers, expected_headers)
-        self.assertFalse(upload.finished)
-        self.assertEqual(upload._chunk_size, _DEFAULT_CHUNKSIZE)
-        self.assertIs(upload._stream, stream)
-        self.assertIsNone(upload._total_bytes)
-        self.assertEqual(upload._content_type, _GENERIC_CONTENT_TYPE)
-        self.assertEqual(upload.resumable_url, resumable_url)
-
-        retry_strategy = upload._retry_strategy
-        self.assertEqual(retry_strategy.max_sleep, 64.0)
-        if num_retries is None:
-            self.assertEqual(retry_strategy.max_cumulative_retry, 600.0)
-            self.assertIsNone(retry_strategy.max_retries)
-        else:
-            self.assertIsNone(retry_strategy.max_cumulative_retry)
-            self.assertEqual(retry_strategy.max_retries, num_retries)
-        self.assertIs(transport, fake_transport)
-        # Make sure we never read from the stream.
-        self.assertEqual(stream.tell(), 0)
-
-        # Check the mocks.
-        request_headers = expected_headers.copy()
-        request_headers["x-upload-content-type"] = _GENERIC_CONTENT_TYPE
-        fake_transport.request.assert_called_once_with(
-            "POST",
-            upload_url,
-            data=json.dumps(metadata).encode("utf-8"),
-            headers=request_headers,
-            timeout=mock.ANY,
-        )
-
-    def test__initiate_resumable_upload(self):
-        self._initiate_resumable_upload_helper()
-
-    def test__initiate_resumable_upload_mtls(self):
-        self._initiate_resumable_upload_helper(mtls=True)
-
-    def test__initiate_resumable_upload_with_retry(self, global_time_lock):
-        self._initiate_resumable_upload_helper(num_retries=11)
-
-    def _do_multipart_upload_success_helper(
-        self, get_boundary, num_retries=None, project=None, mtls=False
-    ):
-        from google.cloud.bigquery.client import _get_upload_headers
-        from google.cloud.bigquery.job import LoadJob
-        from google.cloud.bigquery.job import LoadJobConfig
-        from google.cloud.bigquery.job import SourceFormat
-
-        fake_transport = self._mock_transport(http.client.OK, {})
-        client = self._make_one(project=self.PROJECT, _http=fake_transport)
-        conn = client._connection = make_connection()
-        if mtls:
-            conn.get_api_base_url_for_mtls = mock.Mock(return_value="https://foo.mtls")
-
-        if project is None:
-            project = self.PROJECT
-
-        # Create some mock arguments.
-        data = b"Bzzzz-zap \x00\x01\xf4"
-        stream = io.BytesIO(data)
-        config = LoadJobConfig()
-        config.source_format = SourceFormat.CSV
-        job = LoadJob(None, None, self.TABLE_REF, client, job_config=config)
-        metadata = job.to_api_repr()
-        size = len(data)
-
-        response = client._do_multipart_upload(
-            stream, metadata, size, num_retries, None, project=project
-        )
-
-        # Check the mocks and the returned value.
-        self.assertIs(response, fake_transport.request.return_value)
-        self.assertEqual(stream.tell(), size)
-        get_boundary.assert_called_once_with()
-
-        host_name = "https://foo.mtls" if mtls else "https://bigquery.googleapis.com"
-        upload_url = (
-            f"{host_name}/upload/bigquery/v2/projects/{project}"
-            "/jobs?uploadType=multipart"
-        )
-        payload = (
-            b"--==0==\r\n"
-            b"content-type: application/json; charset=UTF-8\r\n\r\n"
-            b"%(json_metadata)s"
-            b"\r\n"
-            b"--==0==\r\n"
-            b"content-type: */*\r\n\r\n"
-            b"%(data)s"
-            b"\r\n"
-            b"--==0==--"
-        ) % {b"json_metadata": json.dumps(metadata).encode("utf-8"), b"data": data}
-
-        headers = _get_upload_headers(conn.user_agent)
-        headers["content-type"] = b'multipart/related; boundary="==0=="'
-        fake_transport.request.assert_called_once_with(
-            "POST", upload_url, data=payload, headers=headers, timeout=mock.ANY
-        )
-
-    @mock.patch("google.resumable_media._upload.get_boundary", return_value=b"==0==")
-    def test__do_multipart_upload(self, get_boundary):
-        self._do_multipart_upload_success_helper(get_boundary)
-
-    @mock.patch("google.resumable_media._upload.get_boundary", return_value=b"==0==")
-    def test__do_multipart_upload_mtls(self, get_boundary):
-        self._do_multipart_upload_success_helper(get_boundary, mtls=True)
-
-    @mock.patch("google.resumable_media._upload.get_boundary", return_value=b"==0==")
-    def test__do_multipart_upload_with_retry(self, get_boundary, global_time_lock):
-        self._do_multipart_upload_success_helper(get_boundary, num_retries=8)
-
-    @mock.patch("google.resumable_media._upload.get_boundary", return_value=b"==0==")
-    def test__do_multipart_upload_with_custom_project(self, get_boundary):
-        self._do_multipart_upload_success_helper(get_boundary, project="custom-project")
-
     def test_copy_table(self):
         from google.cloud.bigquery.job import CopyJob
 
@@ -5542,143 +5316,6 @@ class TestClient(unittest.TestCase):
                     retry=None,
                     job_retry=None,
                 )
-
-    def test_query_job_rpc_fail_w_conflict_random_id_job_fetch_retries_404(self):
-        """Regression test for https://github.com/googleapis/python-bigquery/issues/2134
-
-        Sometimes after a Conflict, the fetch fails with a 404, but we know
-        because of the conflict that really the job does exist. Retry until we
-        get the job status (or timeout).
-        """
-        job_id = "abc123"
-        creds = _make_credentials()
-        http = object()
-        client = self._make_one(project=self.PROJECT, credentials=creds, _http=http)
-        conn = client._connection = make_connection(
-            # We're mocking QueryJob._begin, so this is only going to be
-            # jobs.get requests and responses.
-            google.api_core.exceptions.TooManyRequests("this is retriable by default"),
-            google.api_core.exceptions.NotFound("we lost your job"),
-            google.api_core.exceptions.NotFound("we lost your job again, sorry"),
-            {
-                "jobReference": {
-                    "projectId": self.PROJECT,
-                    "location": "TESTLOC",
-                    "jobId": job_id,
-                }
-            },
-        )
-
-        job_create_error = google.api_core.exceptions.Conflict("Job already exists.")
-        job_begin_patcher = mock.patch.object(
-            bqjob.QueryJob, "_begin", side_effect=job_create_error
-        )
-        job_id_patcher = mock.patch.object(
-            google.cloud.bigquery._job_helpers,
-            "make_job_id",
-            return_value=job_id,
-        )
-
-        with job_begin_patcher, job_id_patcher:
-            # If get job request fails there does exist a job
-            # with this ID already, retry 404 until we get it (or fails for a
-            # non-retriable reason, see other tests).
-            result = client.query("SELECT 1;", job_id=None)
-
-        jobs_get_path = mock.call(
-            method="GET",
-            path=f"/projects/{self.PROJECT}/jobs/{job_id}",
-            query_params={
-                "projection": "full",
-            },
-            timeout=google.cloud.bigquery.retry.DEFAULT_GET_JOB_TIMEOUT,
-        )
-        conn.api_request.assert_has_calls(
-            # Double-check that it was jobs.get that was called for each of our
-            # mocked responses.
-            [jobs_get_path]
-            * 4,
-        )
-        assert result.job_id == job_id
-
-    def test_query_job_rpc_fail_w_conflict_random_id_job_fetch_retries_404_and_query_job_insert(
-        self, global_time_lock
-    ):
-        """Regression test for https://github.com/googleapis/python-bigquery/issues/2134
-
-        Sometimes after a Conflict, the fetch fails with a 404. If it keeps
-        failing with a 404, assume that the job actually doesn't exist.
-        """
-        job_id_1 = "abc123"
-        job_id_2 = "xyz789"
-        creds = _make_credentials()
-        http = object()
-        client = self._make_one(project=self.PROJECT, credentials=creds, _http=http)
-
-        # We're mocking QueryJob._begin, so that the connection should only get
-        # jobs.get requests.
-        job_create_error = google.api_core.exceptions.Conflict("Job already exists.")
-        job_begin_patcher = mock.patch.object(
-            bqjob.QueryJob, "_begin", side_effect=job_create_error
-        )
-        conn = client._connection = make_connection(
-            google.api_core.exceptions.NotFound("we lost your job again, sorry"),
-            {
-                "jobReference": {
-                    "projectId": self.PROJECT,
-                    "location": "TESTLOC",
-                    "jobId": job_id_2,
-                }
-            },
-        )
-
-        # Choose a small deadline so the 404 retries give up.
-        retry = (
-            google.cloud.bigquery.retry._DEFAULT_GET_JOB_CONFLICT_RETRY.with_deadline(1)
-        )
-        job_id_patcher = mock.patch.object(
-            google.cloud.bigquery._job_helpers,
-            "make_job_id",
-            side_effect=[job_id_1, job_id_2],
-        )
-        retry_patcher = mock.patch.object(
-            google.cloud.bigquery.retry,
-            "_DEFAULT_GET_JOB_CONFLICT_RETRY",
-            retry,
-        )
-
-        with freezegun.freeze_time(
-            "2025-01-01 00:00:00",
-            # 10x the retry deadline to guarantee a timeout.
-            auto_tick_seconds=10,
-        ), job_begin_patcher, job_id_patcher, retry_patcher:
-            # If get job request fails there does exist a job
-            # with this ID already, retry 404 until we get it (or fails for a
-            # non-retriable reason, see other tests).
-            result = client.query("SELECT 1;", job_id=None)
-
-        jobs_get_path_1 = mock.call(
-            method="GET",
-            path=f"/projects/{self.PROJECT}/jobs/{job_id_1}",
-            query_params={
-                "projection": "full",
-            },
-            timeout=google.cloud.bigquery.retry.DEFAULT_GET_JOB_TIMEOUT,
-        )
-        jobs_get_path_2 = mock.call(
-            method="GET",
-            path=f"/projects/{self.PROJECT}/jobs/{job_id_2}",
-            query_params={
-                "projection": "full",
-            },
-            timeout=google.cloud.bigquery.retry.DEFAULT_GET_JOB_TIMEOUT,
-        )
-        conn.api_request.assert_has_calls(
-            # Double-check that it was jobs.get that was called for each of our
-            # mocked responses.
-            [jobs_get_path_1, jobs_get_path_2],
-        )
-        assert result.job_id == job_id_2
 
     def test_query_job_rpc_fail_w_conflict_random_id_job_fetch_succeeds(self):
         from google.api_core.exceptions import Conflict
@@ -10063,181 +9700,6 @@ class TestClientUpload(object):
         )
         transport.request.side_effect = responses
         return transport
-
-    def test__do_resumable_upload(self):
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-        transport = self._make_transport(
-            self._make_resumable_upload_responses(file_obj_len)
-        )
-        client = self._make_client(transport)
-
-        result = client._do_resumable_upload(
-            file_obj, self.EXPECTED_CONFIGURATION, None, None
-        )
-
-        content = result.content.decode("utf-8")
-        assert json.loads(content) == {"size": file_obj_len}
-
-        # Verify that configuration data was passed in with the initial
-        # request.
-        transport.request.assert_any_call(
-            "POST",
-            mock.ANY,
-            data=json.dumps(self.EXPECTED_CONFIGURATION).encode("utf-8"),
-            headers=mock.ANY,
-            timeout=mock.ANY,
-        )
-
-    def test__do_resumable_upload_custom_project(self):
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-        transport = self._make_transport(
-            self._make_resumable_upload_responses(file_obj_len)
-        )
-        client = self._make_client(transport)
-
-        result = client._do_resumable_upload(
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            None,
-            None,
-            project="custom-project",
-        )
-
-        content = result.content.decode("utf-8")
-        assert json.loads(content) == {"size": file_obj_len}
-
-        # Verify that configuration data was passed in with the initial
-        # request.
-        transport.request.assert_any_call(
-            "POST",
-            mock.ANY,
-            data=json.dumps(self.EXPECTED_CONFIGURATION).encode("utf-8"),
-            headers=mock.ANY,
-            timeout=mock.ANY,
-        )
-
-        initiation_url = next(
-            (
-                call[0][1]
-                for call in transport.request.call_args_list
-                if call[0][0] == "POST" and "uploadType=resumable" in call[0][1]
-            ),
-            None,
-        )  # pragma: NO COVER
-
-        assert initiation_url is not None
-        assert "projects/custom-project" in initiation_url
-
-    def test__do_resumable_upload_custom_timeout(self):
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-        transport = self._make_transport(
-            self._make_resumable_upload_responses(file_obj_len)
-        )
-        client = self._make_client(transport)
-
-        client._do_resumable_upload(
-            file_obj, self.EXPECTED_CONFIGURATION, num_retries=0, timeout=3.14
-        )
-
-        # The timeout should be applied to all underlying calls.
-        for call_args in transport.request.call_args_list:
-            assert call_args[1].get("timeout") == 3.14
-
-    def test__do_multipart_upload(self):
-        transport = self._make_transport([self._make_response(http.client.OK)])
-        client = self._make_client(transport)
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-
-        client._do_multipart_upload(
-            file_obj, self.EXPECTED_CONFIGURATION, file_obj_len, None, None
-        )
-
-        # Verify that configuration data was passed in with the initial
-        # request.
-        request_args = transport.request.mock_calls[0][2]
-        request_data = request_args["data"].decode("utf-8")
-        request_headers = request_args["headers"]
-
-        request_content = email.message_from_string(
-            "Content-Type: {}\r\n{}".format(
-                request_headers["content-type"].decode("utf-8"), request_data
-            )
-        )
-
-        # There should be two payloads: the configuration and the binary daya.
-        configuration_data = request_content.get_payload(0).get_payload()
-        binary_data = request_content.get_payload(1).get_payload()
-
-        assert json.loads(configuration_data) == self.EXPECTED_CONFIGURATION
-        assert binary_data.encode("utf-8") == file_obj.getvalue()
-
-    def test__do_multipart_upload_wrong_size(self):
-        client = self._make_client()
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-
-        with pytest.raises(ValueError):
-            client._do_multipart_upload(file_obj, {}, file_obj_len + 1, None, None)
-
-    def test_schema_from_json_with_file_path(self):
-        from google.cloud.bigquery.schema import SchemaField
-
-        file_content = """[
-          {
-            "description": "quarter",
-            "mode": "REQUIRED",
-            "name": "qtr",
-            "type": "STRING"
-          },
-          {
-            "description": "sales representative",
-            "mode": "NULLABLE",
-            "name": "rep",
-            "type": "STRING"
-          },
-          {
-            "description": "total sales",
-            "mode": "NULLABLE",
-            "name": "sales",
-            "type": "FLOAT"
-          }
-        ]"""
-
-        expected = [
-            SchemaField("qtr", "STRING", "REQUIRED", description="quarter"),
-            SchemaField(
-                "rep",
-                "STRING",
-                "NULLABLE",
-                description="sales representative",
-            ),
-            SchemaField(
-                "sales",
-                "FLOAT",
-                "NULLABLE",
-                description="total sales",
-            ),
-        ]
-
-        client = self._make_client()
-        mock_file_path = "/mocked/file.json"
-
-        open_patch = mock.patch(
-            "builtins.open", new=mock.mock_open(read_data=file_content)
-        )
-
-        with open_patch as _mock_file:
-            actual = client.schema_from_json(mock_file_path)
-            _mock_file.assert_called_once_with(mock_file_path)
-            # This assert is to make sure __exit__ is called in the context
-            # manager that opens the file in the function
-            _mock_file().__exit__.assert_called_once()
-
-        assert expected == actual
 
     def test_schema_from_json_with_file_object(self):
         from google.cloud.bigquery.schema import SchemaField
